pyproject.toml
[tool.poetry]
name = "dextbird-sample-bot"
version = "0.1.0"
description = ""
authors = ["mc_fdc <Masato.04.11.2007@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.3.2"
dextbird = {git = "https://github.com/tuna2134/discord-ext-songbird.git"}
aiohttp = "^3.8.5"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
